// backend/src/examples/streaming-fees-modules/interfaces.ts
export interface RecipeDecoderInput {
  transaction: TransactionInfo;
  sharedPoolConnection: RpcPoolConnection;
  craftProgramId: string;
  excludeAccounts: string[];
}

export interface RecipeDecoderOutput {
  decodedRecipe?: {
    kind: 'recipe' | 'process' | 'item';
    data: any;
  };
  craftingMaterial?: string;
  craftingType?: string;
}